+++ homework_06_tests/FILES_CPP/TargetsManager.cpp
@@ -14,7 +14,7 @@ TargetsManager::TargetsManager()
 
 TargetsManager::TargetsManager(const float& arrayTimeStep) : TargetsManager()
 {
-    for (int i = 0; i < targetData.Positions.size(); ++i)
+    for (size_t i = 0; i < targetData.Positions.size(); ++i)
     {
         targets.push_back(TargetController{ targetData.Positions[i], arrayTimeStep });
     }
@@ -22,7 +22,7 @@ TargetsManager::TargetsManager(const float& arrayTimeStep) : TargetsManager()
 
 void TargetsManager::UpdateTargets(const float& simStep)
 {
-    for (int i = 0; i < targets.size(); ++i)
+    for (size_t i = 0; i < targets.size(); ++i)
     {
         targets[i].Update(simStep);
     }
